Revert Executor abstraction and related PRs
* Revert "Move queuing logic from AsynchQueueAgent into Executor (#3683)"
  * This reverts commit 545296609a855e003b4fb353a969ecd8db230e53.
* Revert "Fix null reference exceptions on non started AsynchQueueAgent usage attempt (#3732)"
  * This reverts commit 54c4d02352a13c6b024f5d0fcf2a36d5213ce4ea.
* Revert "Add throw on disposed AsynchAgent usage attempt (#3745)"
  * This reverts commit bc84740dda1d5f6f12314df26f4fef7f7e3d36bd.
