+++ src/Orleans.Core/Messaging/GatewayClientReceiver.cs
@@ -11,7 +11,7 @@ namespace Orleans.Messaging
     /// <summary>
     /// The Receiver class is used by the GatewayConnection to receive messages. It runs its own thread, but it performs all i/o operations synchronously.
     /// </summary>
-    internal class GatewayClientReceiver : SingleTaskAsynchAgent
+    internal class GatewayClientReceiver : AsynchAgent
     {
         private readonly GatewayConnection gatewayConnection;
         private readonly IncomingMessageBuffer buffer;
