+++ src/Orleans.Core/Runtime/AsynchAgent.cs
@@ -15,23 +15,22 @@ namespace Orleans.Runtime
         }
 
         protected readonly ExecutorService executorService;
-        protected IExecutor executor;
         protected CancellationTokenSource Cts;
         protected object Lockable;
         protected ILogger Log;
-        protected readonly string type;
+        private readonly string type;
         protected FaultBehavior OnFault;
-        protected bool disposed;
 
 #if TRACK_DETAILED_STATS
         internal protected ThreadTrackingStatistic threadTracking;
 #endif
 
-        public ThreadState State { get; protected set; }
+        public ThreadState State { get; private set; }
         internal string Name { get; private set; }
 
         protected AsynchAgent(string nameSuffix, ExecutorService executorService, ILoggerFactory loggerFactory)
         {
+            this.executorService = executorService;
             Cts = new CancellationTokenSource();
             var thisType = GetType();
 
@@ -54,7 +53,6 @@ namespace Orleans.Runtime
             OnFault = FaultBehavior.IgnoreFault;
             Log = loggerFactory.CreateLogger(Name);
 
-            this.executorService = executorService;
             AppDomain.CurrentDomain.DomainUnload += CurrentDomain_DomainUnload;
 
 #if TRACK_DETAILED_STATS
@@ -89,7 +87,6 @@ namespace Orleans.Runtime
 
         public virtual void Start()
         {
-            ThrowIfDisposed();
             lock (Lockable)
             {
                 if (State == ThreadState.Running)
@@ -102,29 +99,23 @@ namespace Orleans.Runtime
                     Cts = new CancellationTokenSource();
                 }
 
-                EnsureExecutorInitialized();
-                OnStart();
+                executorService.RunTask(new AsynchAgentTask(() => AgentThreadProc(this), Name));
                 State = ThreadState.Running;
             }
-
             if (Log.IsEnabled(LogLevel.Debug)) Log.Debug("Started asynch agent " + this.Name);
         }
-
-        public virtual void OnStart() { }
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Design", "CA1031:DoNotCatchGeneralExceptionTypes")]
         public virtual void Stop()
         {
             try
             {
-                ThrowIfDisposed();
                 lock (Lockable)
                 {
                     if (State == ThreadState.Running)
                     {
                         State = ThreadState.StopRequested;
                         Cts.Cancel();
-                        executor = null;
                         State = ThreadState.Stopped;
                     }
                 }
@@ -138,6 +129,71 @@ namespace Orleans.Runtime
             }
             Log.Debug("Stopped agent");
         }
+        
+        protected abstract void Run();
+
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Design", "CA1031:DoNotCatchGeneralExceptionTypes")]
+        private static void AgentThreadProc(Object obj)
+        {
+            var agent = obj as AsynchAgent;
+            if (agent == null)
+            {
+                throw new InvalidOperationException("Agent thread started with incorrect parameter type");
+            }
+
+            try
+            {
+                LogStatus(agent.Log, "Starting AsyncAgent {0} on managed thread {1}", agent.Name, Thread.CurrentThread.ManagedThreadId);
+                CounterStatistic.SetOrleansManagedThread(); // do it before using CounterStatistic.
+                CounterStatistic.FindOrCreate(new StatisticName(StatisticNames.RUNTIME_THREADS_ASYNC_AGENT_PERAGENTTYPE, agent.type)).Increment();
+                CounterStatistic.FindOrCreate(StatisticNames.RUNTIME_THREADS_ASYNC_AGENT_TOTAL_THREADS_CREATED).Increment();
+                agent.Run();
+            }
+            catch (Exception exc)
+            {
+                if (agent.State == ThreadState.Running) // If we're stopping, ignore exceptions
+                {
+                    var log = agent.Log;
+                    switch (agent.OnFault)
+                    {
+                        case FaultBehavior.CrashOnFault:
+                            Console.WriteLine(
+                                "The {0} agent has thrown an unhandled exception, {1}. The process will be terminated.",
+                                agent.Name, exc);
+                            log.Error(ErrorCode.Runtime_Error_100023,
+                                "AsynchAgent Run method has thrown an unhandled exception. The process will be terminated.",
+                                exc);
+                            log.Fail(ErrorCode.Runtime_Error_100024, "Terminating process because of an unhandled exception caught in AsynchAgent.Run.");
+                            break;
+                        case FaultBehavior.IgnoreFault:
+                            log.Error(ErrorCode.Runtime_Error_100025, "AsynchAgent Run method has thrown an unhandled exception. The agent will exit.",
+                                exc);
+                            agent.State = ThreadState.Stopped;
+                            break;
+                        case FaultBehavior.RestartOnFault:
+                            log.Error(ErrorCode.Runtime_Error_100026,
+                                "AsynchAgent Run method has thrown an unhandled exception. The agent will be restarted.",
+                                exc);
+                            agent.State = ThreadState.Stopped;
+                            try
+                            {
+                                agent.Start();
+                            }
+                            catch (Exception ex)
+                            {
+                                log.Error(ErrorCode.Runtime_Error_100027, "Unable to restart AsynchAgent", ex);
+                                agent.State = ThreadState.Stopped;
+                            }
+                            break;
+                    }
+                }
+            }
+            finally
+            {
+                CounterStatistic.FindOrCreate(new StatisticName(StatisticNames.RUNTIME_THREADS_ASYNC_AGENT_PERAGENTTYPE, agent.type)).DecrementBy(1);
+                agent.Log.Info(ErrorCode.Runtime_Error_100328, "Stopping AsyncAgent {0} that runs on managed thread {1}", agent.Name, Thread.CurrentThread.ManagedThreadId);
+            }
+        }
 
 #region IDisposable Members
 
@@ -149,15 +205,13 @@ namespace Orleans.Runtime
 
         protected virtual void Dispose(bool disposing)
         {
-            if (!disposing || disposed) return;
+            if (!disposing) return;
 
             if (Cts != null)
             {
                 Cts.Dispose();
                 Cts = null;
             }
-
-            disposed = true;
         }
 
 #endregion
@@ -169,19 +223,17 @@ namespace Orleans.Runtime
 
         internal static bool IsStarting { get; set; }
 
-        private void ThrowIfDisposed()
+        private static void LogStatus(ILogger log, string msg, params object[] args)
         {
-            if (disposed)
+            if (IsStarting)
             {
-                throw new ObjectDisposedException("Cannot access a disposed AsynchAgent"); 
+                // Reduce log noise during silo startup
+                if (log.IsEnabled(LogLevel.Debug)) log.Debug(msg, args);
             }
-        }
-
-        private void EnsureExecutorInitialized()
-        {
-            if (executor == null)
+            else
             {
-                executor = executorService.GetExecutor(new GetExecutorRequest(GetType(), Name, Cts));
+                // Changes in agent threads during all operations aside for initial creation are usually important diag events.
+                log.Info(msg, args);
             }
         }
     }
