+++ src/Orleans.Core/Runtime/AsynchQueueAgent.cs
@@ -1,34 +1,157 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Threading;
 using Microsoft.Extensions.Logging;
 using Orleans.Runtime.Configuration;
 
 namespace Orleans.Runtime
 {
-    internal abstract class AsynchQueueAgent<T> : AsynchAgent where T : IOutgoingMessage
+    internal abstract class AsynchQueueAgent<T> : AsynchAgent, IDisposable where T : IOutgoingMessage
     {
+        private BlockingCollection<T> requestQueue;
+        private QueueTrackingStatistic queueTracking;
+
         protected AsynchQueueAgent(string nameSuffix, ExecutorService executorService, ILoggerFactory loggerFactory)
             : base(nameSuffix, executorService, loggerFactory)
         {
-            ProcessAction = state => Process((T)state);
+            requestQueue = new BlockingCollection<T>();
+            if (StatisticsCollector.CollectQueueStats)
+            {
+                queueTracking = new QueueTrackingStatistic(base.Name);
+            }
         }
-
-        public WaitCallback ProcessAction { get; }
 
         public void QueueRequest(T request)
         {
-            if (State != ThreadState.Running)
+            if (requestQueue==null)
             {
-                throw new InvalidOperationException("Not running agent usage attempt");
+                return;
             }
 
-            executor.QueueWorkItem(ProcessAction, request);
+#if TRACK_DETAILED_STATS
+            if (StatisticsCollector.CollectQueueStats)
+            {
+                queueTracking.OnEnQueueRequest(1, requestQueue.Count, request);
+            }
+#endif
+
+            requestQueue.Add(request);
         }
 
         protected abstract void Process(T request);
-        
-        public int Count => executor.WorkQueueCount;
+
+        protected override void Run()
+        {
+#if TRACK_DETAILED_STATS
+            if (StatisticsCollector.CollectThreadTimeTrackingStats)
+            {
+                threadTracking.OnStartExecution();
+                queueTracking.OnStartExecution();
+            }
+#endif
+            try
+            {
+                RunNonBatching();
+            }
+            finally
+            {
+#if TRACK_DETAILED_STATS
+                if (StatisticsCollector.CollectThreadTimeTrackingStats)
+                {
+                    threadTracking.OnStopExecution();
+                    queueTracking.OnStopExecution();
+                }
+#endif
+            }
+        }
+
+
+        protected void RunNonBatching()
+        {            
+            while (true)
+            {
+                if (Cts == null || Cts.IsCancellationRequested)
+                {
+                    return;
+                }
+                T request;
+                try
+                {
+                    request = requestQueue.Take();
+                }
+                catch (InvalidOperationException)
+                {
+                    Log.Info(ErrorCode.Runtime_Error_100312, "Stop request processed");
+                    break;
+                }
+#if TRACK_DETAILED_STATS
+                if (StatisticsCollector.CollectQueueStats)
+                {
+                    queueTracking.OnDeQueueRequest(request);
+                }
+                if (StatisticsCollector.CollectThreadTimeTrackingStats)
+                {
+                    threadTracking.OnStartProcessing();
+                }
+#endif
+                Process(request);
+#if TRACK_DETAILED_STATS
+                if (StatisticsCollector.CollectThreadTimeTrackingStats)
+                {
+                    threadTracking.OnStopProcessing();
+                    threadTracking.IncrementNumberOfProcessed();
+                }
+#endif
+            }
+        }
+
+        public override void Stop()
+        {
+#if TRACK_DETAILED_STATS
+            if (StatisticsCollector.CollectThreadTimeTrackingStats)
+            {
+                threadTracking.OnStopExecution();
+            }
+#endif
+            requestQueue?.CompleteAdding();
+            base.Stop();
+        }
+
+        protected void DrainQueue(Action<T> action)
+        {
+            T request;
+            while (requestQueue.TryTake(out request))
+            {
+                action(request);
+            }
+        }
+
+        public virtual int Count
+        {
+            get
+            {
+                return requestQueue.Count;
+            }
+        }
+
+        #region IDisposable Members
+
+        protected override void Dispose(bool disposing)
+        {
+            if (!disposing) return;
+            
+#if TRACK_DETAILED_STATS
+            if (StatisticsCollector.CollectThreadTimeTrackingStats)
+            {
+                threadTracking.OnStopExecution();
+            }
+#endif
+            base.Dispose(disposing);
+
+            requestQueue?.Dispose();
+            requestQueue = null;
+        }
+
+        #endregion
     }
 }
