+++ src/Orleans.Runtime/GrainDirectory/MultiClusterRegistration/GlobalSingleInstanceActivationMaintainer.cs
@@ -14,7 +14,7 @@ using Orleans.Runtime.MultiClusterNetwork;
 
 namespace Orleans.Runtime.GrainDirectory
 {
-    internal class GlobalSingleInstanceActivationMaintainer : SingleTaskAsynchAgent
+    internal class GlobalSingleInstanceActivationMaintainer : AsynchAgent
     {
         private readonly object lockable = new object();
         private readonly LocalGrainDirectory router;
