+++ src/Orleans.Runtime/Messaging/IncomingMessageAgent.cs
@@ -5,7 +5,7 @@ using Orleans.Runtime.Scheduler;
 
 namespace Orleans.Runtime.Messaging
 {
-    internal class IncomingMessageAgent : SingleTaskAsynchAgent
+    internal class IncomingMessageAgent : AsynchAgent
     {
         private readonly IMessageCenter messageCenter;
         private readonly ActivationDirectory directory;
@@ -42,30 +42,48 @@ namespace Orleans.Runtime.Messaging
 
         protected override void Run()
         {
-            CancellationToken ct = Cts.Token;
-            while (true)
+            try
             {
-                // Get an application message
-                var msg = messageCenter.WaitMessage(category, ct);
-                if (msg == null)
-                {
-                    if (Log.IsEnabled(LogLevel.Debug)) Log.Debug("Dequeued a null message, exiting");
-                    // Null return means cancelled
-                    break;
-                }
-
 #if TRACK_DETAILED_STATS
                 if (StatisticsCollector.CollectThreadTimeTrackingStats)
                 {
-                    threadTracking.OnStartProcessing();
+                    threadTracking.OnStartExecution();
                 }
 #endif
-                ReceiveMessage(msg);
+                CancellationToken ct = Cts.Token;
+                while (true)
+                {
+                    // Get an application message
+                    var msg = messageCenter.WaitMessage(category, ct);
+                    if (msg == null)
+                    {
+                        if (Log.IsEnabled(LogLevel.Debug)) Log.Debug("Dequeued a null message, exiting");
+                        // Null return means cancelled
+                        break;
+                    }
+
+ #if TRACK_DETAILED_STATS
+                    if (StatisticsCollector.CollectThreadTimeTrackingStats)
+                    {
+                        threadTracking.OnStartProcessing();
+                    }
+ #endif
+                    ReceiveMessage(msg);
+ #if TRACK_DETAILED_STATS
+                    if (StatisticsCollector.CollectThreadTimeTrackingStats)
+                    {
+                        threadTracking.OnStopProcessing();
+                        threadTracking.IncrementNumberOfProcessed();
+                    }
+ #endif
+                }
+            }
+            finally
+            {
 #if TRACK_DETAILED_STATS
                 if (StatisticsCollector.CollectThreadTimeTrackingStats)
                 {
-                    threadTracking.OnStopProcessing();
-                    threadTracking.IncrementNumberOfProcessed();
+                    threadTracking.OnStopExecution();
                 }
 #endif
             }
