+++ src/Orleans.Runtime/Silo/Watchdog.cs
@@ -3,9 +3,11 @@ using System.Collections.Generic;
 using System.Threading;
 using Microsoft.Extensions.Logging;
 
+
 namespace Orleans.Runtime
 {
-    internal class Watchdog : SingleTaskAsynchAgent
+
+    internal class Watchdog : AsynchAgent
     {
         private static readonly TimeSpan heartbeatPeriod = TimeSpan.FromMilliseconds(1000);
         private readonly TimeSpan healthCheckPeriod;
